+++ src/odb.c
@@ -164,11 +164,11 @@ int git_odb__hashlink(git_oid *out, const char *path)
 		char *link_data;
 		ssize_t read_len;
 
-		link_data = git__malloc(size);
+		link_data = git__malloc((size_t)size);
 		if (link_data == NULL)
 			return GIT_ENOMEM;
 
-		read_len = p_readlink(path, link_data, size + 1);
+		read_len = p_readlink(path, link_data, (size_t)(size + 1));
 		if (read_len != (ssize_t)size)
 			return git__throw(GIT_EOSERR, "Failed to read symlink data");
